Move back non integration tests back into module
