+++ src/error/mod.rs
@@ -1,6 +1,3 @@
 mod app_error;
 
-#[cfg(test)]
-mod app_error_tests;
-
 pub use app_error::{AppError, AppResponse};
